BIG OL' UPDATE
This rips the scheduler out from the app. Starts work on #91.

Fixes #94, #136, #137, #139
