+++ contrib/uwsgi/hourglass-uwsgi.ini
@@ -1,6 +1,7 @@
 [uwsgi]
 chdir = /opt/hourglass
-wsgi-file = wsgi.py
+wsgi-file = contrib/uwsgi/wsgi.py
+mule = contrib/uwsgi/scheduler.py
 callable = app
 master = true
 socket = /var/run/hourglass.sock
