+++ hourglass/scheduler.py
@@ -1,22 +1,24 @@
 import asyncio
 import importlib
-from hourglass.backends import db
+from hourglass.db import db
 from hourglass.backends.cache import Zone
+from hourglass.app import create_app
+from flask import current_app
 
 
-class UwsgiScheduler(object):
+class Scheduler(object):
 
-    def __init__(self, app):
-        self.app = app
+    def __init__(self, config_file):
         self.loop = asyncio.get_event_loop()
-        self.interval = int(app.config['hourglass'].get(
-            'scheduler_interval', 30))
-        # load the modules so sqlalchemy knows about them
-        self._load_zones_modules()
+        self.config_file = config_file
+        with create_app(self.config_file).app_context():
+            self.config = current_app.config
+            self.interval = int(current_app.config['hourglass'].get(
+                'scheduler_interval', 30))
 
     def _load_zones(self):
         zones = []
-        for name, config in self.app.config['zones'].items():
+        for name, config in self.config['zones'].items():
             backend = config.get('backend')
             package = backend.split(':')[0]
             class_name = backend.split(':')[1]
@@ -25,30 +27,23 @@ class UwsgiScheduler(object):
             zones.append(zone_class(name, **config))
         return zones
 
-    def _load_zones_modules(self):
-        for name, config in self.app.config['zones'].items():
-            backend = config.get('backend')
-            package = backend.split(':')[0]
-            importlib.import_module(package)
-
     def create_cache_db(self):
-        with self.app.app_context():
+        self._load_zones()
+        with create_app(self.config_file).app_context():
+            current_app.logger.info('Creating cache database')
             db.drop_all(bind='cache')
             db.create_all(bind='cache')
             zones = self._load_zones()
             db.session.bulk_save_objects(zones)
             db.session.commit()
-        self.run_tasks()
 
-    def run_tasks(self, kill_loop=False):
-        with self.app.app_context():
+    def run_tasks(self):
+        with create_app(self.config_file).app_context():
             tasks = []
             for zone in Zone.query.all():
-                tasks.extend(zone.get_update_tasks(self.app))
+                tasks.extend(zone.get_update_tasks(current_app))
             results = self.loop.run_until_complete(asyncio.gather(*tasks))
             for result in results:
                 db.session.bulk_save_objects(result)
             db.session.commit()
-            self.app.logger.info('Cache database updated')
-        if kill_loop:
-            self.loop.close()
+            current_app.logger.info('Cache database updated')
