+++ setup.cfg
@@ -1,6 +1,5 @@
 [metadata]
 name = hourglass
-version = 0.1.0
 summary = A multi-user/group web dashboard for Sensu
 description-file =
     README.md
@@ -24,4 +23,4 @@ universal = 1
 
 [entry_points]
 console_scripts = 
-    hourglass = hourglass.shell:run_hourglass
+    hourglass = hourglass.shell:main
